- Login UI done - Sign Up UI done - Few adjusments and enhancements around the app in the profile etc.. - Exercises for physical done

+++ app/src/main/java/com/example/wellnessfusionapp/Models/ExerciseLog.kt
@@ -17,8 +17,8 @@ data class ExerciseDetail(
     val exerciseName: String,
     val sets: Int = 0,
     val reps: Int = 0,
-    val weight: Float = 0f
+    val weight: Int = 0
 )
 {
-    constructor() : this("", "", 0, 0, 0f)
+    constructor() : this("", "", 0, 0, 0)
 }
